requirements.txt
flask>=2.0.0
openai>=1.50.0
python-dotenv>=1.0.0
Pillow>=9.0.0
hypothesis>=6.0.0
pytest>=7.0.0
